Refactored code for a consistent naming convention
