+++ app/src/main/java/com/litlgroup/litl/utils/SpacesItemDecoration.java
@@ -1,4 +1,4 @@
-package com.litlgroup.litl.helpers;
+package com.litlgroup.litl.utils;
 
 import android.graphics.Rect;
 import android.support.v7.widget.RecyclerView;
